@@ -3,14 +3,17 @@ name = "hackaton-ml"
 version = "0.1.0"
 description = "Add your description here"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 dependencies = [
     "aiohttp>=3.10.11",
+    "editdistance>=0.8.1",
+    "matplotlib>=3.7.5",
     "numpy==1.22.4",
-    "opencv-python==3.4.8.29",
+    "opencv-python>=4.5.0",
     "requests>=2.32.4",
-    "scipy==1.8.1",
-    "shapely==1.8.2",
-    "torch==1.4",
-    "torchvision==0.5.0",
+    "scipy>=1.8.1",
+    "shapely>=1.8.2",
+    "tqdm>=4.67.1",
+    "torch>=1.9.0",
+    "torchvision>=0.10.0",
 ]
